refactor: split pure task-naming and URL utilities out of task-utils
