+++ src/actions/task-utils.ts
@@ -7,46 +7,6 @@ import type { Logger } from "../infra/logger";
 
 const DEFAULT_MAX_RETRIES = 5;
 
-const MAX_TASK_NAME_LENGTH = 32;
-
-export function generateTaskName(
-	prefix: string,
-	repo: string,
-	issueNumber: number,
-): string {
-	const issueStr = String(issueNumber);
-	// Format: {prefix}-{repo}-{issueNumber}
-	// Coder API enforces a 32-character limit on task names.
-	const overhead = prefix.length + 1 + 1 + issueStr.length; // prefix + "-" + "-" + issueNumber
-	const maxRepoLength = MAX_TASK_NAME_LENGTH - overhead;
-	if (maxRepoLength <= 0) {
-		throw new Error(
-			`Task name prefix "${prefix}" and issue number ${issueNumber} leave no room for the repo name (max ${MAX_TASK_NAME_LENGTH} chars)`,
-		);
-	}
-	const truncatedRepo =
-		repo.length > maxRepoLength
-			? repo.slice(0, maxRepoLength).replace(/-+$/, "")
-			: repo;
-	return `${prefix}-${truncatedRepo}-${issueStr}`;
-}
-
-export function parseIssueURL(url: string): {
-	owner: string;
-	repo: string;
-	issueNumber: number;
-} {
-	const match = url.match(/github\.com\/([^/]+)\/([^/]+)\/issues\/(\d+)/);
-	if (!match) {
-		throw new Error(`Invalid GitHub issue URL: ${url}`);
-	}
-	return {
-		owner: match[1],
-		repo: match[2],
-		issueNumber: Number.parseInt(match[3], 10),
-	};
-}
-
 export async function lookupAndEnsureActiveTask(
 	coder: CoderClient,
 	coderUsername: string | undefined,
